fix a few things

+++ pypp/pypp_shutil.h
@@ -11,16 +11,16 @@ namespace shutil {
 inline void rmtree(const PyStr &path) {
     std::error_code ec;
     std::filesystem::path fs_path(path.str());
+    if (!std::filesystem::exists(fs_path)) {
+        throw FileNotFoundError(
+            PyStr("The system cannot find the file specified: ") + path);
+    }
+    if (!std::filesystem::is_directory(fs_path)) {
+        throw NotADirectoryError(PyStr("The directory name is invalid: ") +
+                                 path);
+    }
     std::filesystem::remove_all(fs_path, ec);
     if (ec) {
-        if (ec == std::errc::no_such_file_or_directory) {
-            throw FileNotFoundError(
-                PyStr("The system cannot find the file specified: ") + path);
-        }
-        if (ec == std::errc::not_a_directory) {
-            throw NotADirectoryError(PyStr("The directory name is invalid: ") +
-                                     path);
-        }
         if (ec == std::errc::permission_denied) {
             throw PermissionError(PyStr("Permission denied: ") + path);
         }
